[PiranhaJava] Switch config to structured properties.json (#39)
The config file is moved to `properties.json`, removing the necessity for `piranha.properties` and allowing for more extensible configuration. This PR 
uses that flexibility to add support for specifying the return type, receiver type, 
and flag argument index from each experimentation API method configured.

This is an enhancement and a fix for #38 and #6 . 

See included documentation for the full format of the json configuration file.