[PiranhaJava] Fix issue with code not being deleted across files (#99)
This fixes an issue where the end position of overlaps, as calculated
by `overLaps(...)` was sometimes not being cleared from one compilation
unit (java file) to the next. This resulted in refactorings being
skipped for the first X lines of the second (and third, and so on)
Java file being processed, where X is the last line of the the
last refactoring of the first file.

In particular, for files with similar uses of flags, this resulted
in only the uses within the first file being analyzed being properly
cleaned.

The fix is simply to reset `endPos` at the beginning of any new
compilation unit. This PR additionally adds a regression unit
test for the same issue.