Initial code and test scenarios for different api style (#163)
This PR adds functionality for matching experiment APIs that require invoking a chain of method invocations.
This functionality can be toggled by setting the allow_method_chain in the clean_up options to true.
In addition, it also tests the possibility of using Piranha for experiment APIs (i) with non pre-determined names and (ii) non pre-determined membership to the treatment/control group.
The PR contains many tests for a mock API that requires invoking a chain of method invocations. The tests in this PR transform the input properties by replacing the tokens [notIsTreated] and [flagName] with values based on arguments passed to Piranha.